refactor(BA-6510): consolidate merged-view types into app_config_fragment
- move AppConfigData into data/app_config_fragment/types.py and type its
  user_id as UserID
- rename AppConfigSearchResult -> ScopedAppConfigSearchResult and move it
  alongside (scoped/admin merged search both return it)
- drop the now-empty data/app_config module
