--- src/ai/backend/manager/data/app_config/types.py
@@ -1,33 +1,0 @@
-from __future__ import annotations
-
-from collections.abc import Mapping, Sequence
-from dataclasses import dataclass
-from typing import Any
-from uuid import UUID
-
-from ai.backend.manager.data.app_config_fragment.types import AppConfigFragmentData
-
-
-@dataclass(frozen=True)
-class AppConfigData:
-    """Service-layer return type for the merged AppConfig view.
-
-    `fragments` are ordered low → high merge priority (by fragment
-    `rank`). `config` is the deep-merged result, projected to `None`
-    when every contributing fragment is empty.
-    """
-
-    user_id: UUID
-    name: str
-    fragments: Sequence[AppConfigFragmentData]
-    config: Mapping[str, Any] | None
-
-
-@dataclass(frozen=True)
-class AppConfigSearchResult:
-    """Result from searching merged `AppConfig` views."""
-
-    items: list[AppConfigData]
-    total_count: int
-    has_next_page: bool
-    has_previous_page: bool
